Game saving code extracted
